+++ src/app/[locale]/page.tsx
@@ -69,6 +69,10 @@ export async function generateMetadata({
         seo.languageAlternates,
     },
     robots: seo.robots,
+    openGraph:
+      seo.openGraph,
+    twitter:
+      seo.twitter,
   };
 }