+++ src/app/[locale]/products/[slug]/page.tsx
@@ -93,6 +93,10 @@ export async function generateMetadata({
         seo.languageAlternates,
     },
     robots: seo.robots,
+    openGraph:
+      seo.openGraph,
+    twitter:
+      seo.twitter,
   };
 }